+++ circus-train-s3-s3-copier/src/test/java/com/hotels/bdp/circustrain/s3s3copier/S3S3CopierOptionsTest.java
@@ -115,5 +115,18 @@ public class S3S3CopierOptionsTest {
     S3S3CopierOptions options = new S3S3CopierOptions(copierOptions);
     assertThat(options.getMaxCopyAttempts(), is(3));
   }
+  
+  @Test
+  public void getAssumedRole() throws Exception {
+    copierOptions.put(S3S3CopierOptions.Keys.ASSUME_ROLE.keyName(), "iam:role:1234:user");
+    S3S3CopierOptions options = new S3S3CopierOptions(copierOptions);
+    assertThat(options.getAssumedRole(), is("iam:role:1234:user"));
+  }
+  
+  @Test
+  public void getAssumedRoleDefaultIsNull() throws Exception {
+    S3S3CopierOptions options = new S3S3CopierOptions(copierOptions);
+    assertNull(options.getAssumedRole());
+  }
 
 }