+++ circus-train-s3-s3-copier/src/test/java/com/hotels/bdp/circustrain/s3s3copier/S3S3CopierTest.java
@@ -103,7 +103,8 @@ public class S3S3CopierTest {
     client.createBucket("source");
     client.createBucket("target");
 
-    when(s3ClientFactory.newInstance(any(AmazonS3URI.class), any(S3S3CopierOptions.class))).thenReturn(newClient());
+    when(s3ClientFactory.newInstance(any(AmazonS3URI.class), any(S3S3CopierOptions.class)))
+        .thenReturn(newClient());
   }
 
   private AmazonS3 newClient() {